Clean, refactor, and speed up AutoComplete
