+++ src/AutoCompleteText.css
@@ -3,7 +3,6 @@
 	margin: 8px 0;
     font-size: 14px;
     background-color: #eee;
-    border: 1px solid grey;
     box-sizing: border-box;
 }
